move fp8 init to its own file
